Improved functionality
Found some minor bugs like sometimes the agent destination is a non walkable area, so i now check if the destination is actual a place we can walk within in bounds of the map. Also we now dynamically find the distance from the center of the square floor to the edge so I can change the size of the map if I wanted to. Implemented abstraction to the chasing code. This code is now in the Chase(float distance) method instead of in naked in the Update() method.