Player can shoot bullets in forward direction
I added a game manager so that i can manager when a player and enemy shoots... now thinking about it, i should have just created a parent class for this. Anyways! The player can shoot in the forward direction, and now there are boundaries around the map so the player cannot fall off. Created basic bullet behavior