Game Over sequence completed
Player now receives a game over screen when they run out of health. They can no longer shoot or move in this state and enemies cannot shoot either but can move. The player has the ability to restart the game. But, in the process of restarting the game, I had to take away the PlayerManager because it was causing errors. So in the next push, I will try to reimplement the PlayerManager because it helps save data