Minor Bug fixed
Sometimes when the enemy that shot the bullet is destroyed before bulletBehaviour can destroy the bullet, an error would appear about references. Now that does not happen anymore. There is no need to keep persistant data about the player or floor because it causes issues. This should only be for scores, settings, etc and not gameplay mechanics