Created Boss Prefab
the boss is made, but it shoots weird. I have to figure something out with the ShootBullet() and Attack() method because I need to change just one small part of it to adjust the height that the bullet shoots from. Maybe even need to change the bullet physics for the boss