HowToPlay panel active & QUIT button
The how to play button now leads to a different panel to tell the player what they need to do when they are playing. It also has a return button so the player can go back to the main menu. Also, forgot to add a quit button, so I added that and gave its functionality