(bugs remain) creating leaderboard
In the process of creating the leaderboard. There is one known issue that we cannot add a score that is smaller than all the other scores but im kinda tired so i dont want to fix that right now. Created a new class called Score to keep track of name and score of the current player which helps create the leaderboard