highscore is secure
The highscore now shows up at the main menu stats panel and if there is no highscore it displays "none". The same logic is used for the game over panel because we display the highscore there too

The next thing to do is creat a nice skydome and make this data persistent through sessions. Then, I will be happy with the game