cleaning code - boss bug found
while i was cleaning up the code and making things look nice i noticed that that boss has a bug. If you run around it enough, it will stop facing the player and that is an issue because the enemy always needs to face the player even if you do run around it 100 times. So now I have to figure out how to solve this bug