fixed a bit
So the boss does not get stuck in one position now but its rotation still is slightly off when you run around the boss and then run away. However, after running away from it for a while, the boss's rotation is good again. So this bug is probably something to do with the AI calculations that I am unaware of. We will leave it as it is now